requirements.txt
# ============================== Core Dependencies ==============================
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
python-dotenv==1.1.1
psutil==7.0.0
PyJWT==2.10.1
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.14.1
typing-inspection==0.4.1

# ============================ Security & Cryptography ============================
cryptography==45.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
cffi==1.17.1
pycparser==2.22
rsa==4.9.1

# ============================== Networking & HTTP ==============================
requests==2.32.4
urllib3==2.5.0
httpx==0.28.1
httpcore==1.0.9
httplib2==0.22.0
CacheControl==0.14.3
cachetools==5.5.2
idna==3.10
charset-normalizer==3.4.2
certifi==2025.7.14
sniffio==1.3.1

# ================================ Async / I/O =================================
anyio==4.9.0
websockets==14.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0

# ===================== Google Cloud & Vertex AI SDKs ==========================
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
google-cloud-core==2.4.3
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.58.0
google-cloud-firestore==2.21.0
google-cloud-bigquery==3.35.0
google-cloud-resource-manager==1.14.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.62.3
proto-plus==1.26.1
protobuf==4.25.8
google-genai==1.3.0

# ============================ Firebase Admin SDK ==============================
firebase-admin==6.4.0

# ================================ ML / Math ==================================
numpy==2.3.1
shapely==2.1.1

# =========================== Dev / Utils / Parsing ============================
msgpack==1.1.1
python-dateutil==2.9.0.post0
pyparsing==3.2.3
docstring_parser==0.16
python-multipart==0.0.20
pandas
numpy